observe memo.

+++ app/src/main/java/dev/marshi/jetmemo/data/database/entity/MemoEntity.kt
@@ -8,16 +8,16 @@ import dev.marshi.jetmemo.domain.entity.MemoId
 
 @Entity(tableName = "memo")
 data class MemoEntity(
-    @PrimaryKey(autoGenerate = true) val id: Int,
+    @PrimaryKey(autoGenerate = true) val id: Long,
     @ColumnInfo(name = "text") val text: String?,
     @ColumnInfo(name = "inserted_at") val insertedAt: Long,
     @ColumnInfo(name = "updated_at") val updatedAt: Long,
 ) {
     companion object {
-        fun from(id: Int? = null, text: String): MemoEntity {
+        fun from(id: Long? = null, text: String): MemoEntity {
             val now = System.currentTimeMillis()
             return MemoEntity(
-                id = id ?: 0,
+                id = id ?: 0L,
                 text = text,
                 insertedAt = now,
                 updatedAt = now,
