+++ app/src/main/java/dev/marshi/jetmemo/domain/entity/MemoId.kt
@@ -2,10 +2,10 @@ package dev.marshi.jetmemo.domain.entity
 
 @JvmInline
 value class MemoId private constructor(
-    val value: Int
+    val value: Long
 ) {
     companion object {
-        fun from(value: Int): MemoId {
+        fun from(value: Long): MemoId {
             return MemoId(value)
         }
     }
